requirements.txt
argcomplete==1.9.5
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
Beaker==1.10.0
bleach==3.1.0
blivet==3.1.2
blivet-gui==2.1.10
boto3==1.9.164
botocore==1.12.164
Brlapi==0.6.7
cffi==1.11.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
conda==4.5.13
cookies==2.2.1
coverage==4.5.1
cryptography==2.3
cupshelpers==1.0
cycler==0.10.0
cytoolz==0.9.0.1
decorator==4.3.0
defusedxml==0.5.0
distro==1.3.0
Django==2.2.1
django-crispy-forms==1.7.2
django-storages==1.7.1
docutils==0.14
entrypoints==0.3
fros==1.1
funcsigs==1.0.2
gpg==1.12.0
graphviz==0.10.1
gunicorn==19.9.0
humanize==0.5.1
idna==2.7
iniparse==0.4
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.1.0
ipywidgets==7.4.2
isc==2.0
isort==4.3.19
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
langtable==0.0.41
lazy-object-proxy==1.4.1
libvirt-python==4.7.0
Mako==1.0.9.dev0
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
ntplib==0.3.3
numpy==1.16.2
olefile==0.46
ordered-set==2.0.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
Paste==2.0.3
pbr==4.1.1
pexpect==4.7.0
pickleshare==0.7.5
pid==2.2.3
Pillow==5.3.0
ply==3.9
productmd==1.21
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pwquality==1.4.0
pycairo==1.17.1
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.72
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
pykickstart==3.16
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyparted==3.11.0
pyrsistent==0.14.11
PySocks==1.6.8
python-augeas==0.5.0
python-dateutil==2.8.0
python-meh==0.47
pytz==2018.5
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.20.0
requests-file==1.4.3
requests-ftp==0.3.1
responses==0.9.0
rpm==4.14.2.1
ruamel.yaml==0.15.41
s3transfer==0.2.1
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
sepolicy==1.1
setools==4.1.1
simpleline==1.4
six==1.11.0
sklearn==0.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.7
sqlparse==0.3.0
SSSDConfig==2.1.0
systemd-python==234
Tempita==0.5.1
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
xlrd==1.2.0
